Update: hide collapsed panes and status bar border.
Collapsed panes before had some annoying borders that looked out of place along with the border of the status bar. Changed that theme so that those borders will not be visible by default from now on. However, If the user intends to have them a `style-setting` is also added to toggle between the two states.